Send via `PHPMailer` class
